requirements.txt
fastapi[all]==0.116.1
python-pptx==1.0.2
google-generativeai==0.8.5
python-dotenv==1.1.1
slowapi==0.1.9
httpx